// src/api/generated/schemas/notificationChannelType.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */

export type NotificationChannelType =
  (typeof NotificationChannelType)[keyof typeof NotificationChannelType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NotificationChannelType = {
  Email: 'email',
  Sms: 'sms',
  Push: 'push',
  Internal: 'internal',
} as const;